GW-98: Refactored folder structure for games
