+++ client/src/pages/games/pages/dashboard/components/Points.jsx
