+++ client/src/pages/games/pages/onboarding/Onboarding.jsx
